+++ bit_manipulations.hpp
@@ -214,6 +214,6 @@ int stone_sum(board);
 int bit_to_pos(uint64_t);
 uint64_t rotr(uint64_t, int);
 
-uint16_t toBase3(uint8_t, uint8_t);
+uint16_t toBase3(uint16_t, uint16_t);
 
 } // namespace bit_manipulations
